# docs/api/books.openapi.yaml
openapi: 3.1.0
info:
  title: Tako API - Books Domain
  version: 1.0.0
  description: |
    API pour la recherche et la consultation de livres via différentes sources.
    
    ## Providers disponibles
    
    | Provider | Description | Clé API | Rate Limit |
    |----------|-------------|---------|------------|
    | **Google Books** | Plus grande base de données de livres numériques | Requise | 1000 req/jour (gratuit) |
    | **Open Library** | Bibliothèque ouverte et collaborative | Non | Non spécifié |
    
    ## Fonctionnalités
    
    - Recherche par titre, auteur, ISBN
    - Détails complets (synopsis, couvertures, métadonnées)
    - Support multi-langues
    - Validation et conversion ISBN (ISBN-10/ISBN-13)
  contact:
    name: Tako API
    url: https://github.com/tako-api
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/books
    description: Base path pour le domaine books
...
paths:
...
  /googlebooks/search:
    get:
      tags: [Google Books]
      summary: Rechercher des livres
      description: |
        Recherche de livres par titre ou ISBN via Google Books.
        
        La recherche détecte automatiquement si la requête est un ISBN
        et ajuste le type de recherche en conséquence.
      operationId: searchGoogleBooks
      parameters:
        - name: q
          in: query
          required: true
          description: Terme de recherche (titre ou ISBN)
          schema:
            type: string
            example: "Harry Potter"
        - name: maxResults
          in: query
          description: Nombre maximum de résultats (1-40)
          schema:
            type: integer
            minimum: 1
            maximum: 40
            default: 20
        - name: orderBy
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum: [relevance, newest]
            default: relevance
        - name: langRestrict
          in: query
          description: Restreindre à une langue (code ISO 2 lettres)
          schema:
            type: string
            example: "fr"
      responses:
        '200':
          description: Résultats de recherche
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSearchResponse'
        '400':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /googlebooks/search/author:
    get:
      tags: [Google Books]
      summary: Rechercher par auteur
      description: Recherche de livres par nom d'auteur via Google Books
      operationId: searchGoogleBooksByAuthor
      parameters:
        - name: author
          in: query
          required: true
          description: Nom de l'auteur
          schema:
            type: string
            example: "J.K. Rowling"
        - name: maxResults
          in: query
          description: Nombre maximum de résultats (1-40)
          schema:
            type: integer
            minimum: 1
            maximum: 40
            default: 20
        - name: orderBy
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum: [relevance, newest]
            default: relevance
        - name: langRestrict
          in: query
          description: Restreindre à une langue
          schema:
            type: string
      responses:
        '200':
          description: Résultats de recherche
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSearchResponse'
        '400':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /googlebooks/{volumeId}:
    get:
      tags: [Google Books]
      summary: Détails d'un livre
      description: Récupère les détails complets d'un livre par son ID Google Books
      operationId: getGoogleBookById
      parameters:
        - name: volumeId
          in: path
          required: true
          description: ID du volume Google Books
          schema:
            type: string
            example: "zyTCAlFPjgYC"
      responses:
        '200':
          description: Détails du livre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailResponse'
        '404':
          description: Livre non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ═══════════════════════════════════════════════════════════════════════════
  # OPEN LIBRARY
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  schemas:
...
    BookSearchResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            source:
              type: string
              enum: [googlebooks, openlibrary]
            query:
              type: string
            searchType:
              type: string
              enum: [text, isbn, author]
            total:
              type: integer
            returned:
              type: integer
            pagination:
              type: object
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalResults:
                  type: integer
                hasMore:
                  type: boolean
            timestamp:
              type: string
              format: date-time
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookSummary'
    BookSummary:
      type: object
      properties:
        id:
          type: string
          description: ID unique du livre (provider-specific)
        type:
          type: string
          example: "book"
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        authors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              role:
                type: string
                default: "author"
        publisher:
          type: string
          nullable: true
        publishedDate:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        categories:
          type: array
          items:
            type: string
        language:
          type: string
          nullable: true
        identifiers:
          type: object
          properties:
            googlebooks:
              type: string
              nullable: true
            openlibrary:
              type: string
              nullable: true
            isbn:
              type: string
              nullable: true
            isbn10:
              type: string
              nullable: true
            isbn13:
              type: string
              nullable: true
        covers:
          type: object
          properties:
            large:
              type: string
              format: uri
              nullable: true
            medium:
              type: string
              format: uri
              nullable: true
            small:
              type: string
              format: uri
              nullable: true
        description:
          type: string
          nullable: true
        pageCount:
          type: integer
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        source:
          type: string
          enum: [googlebooks, openlibrary]
    BookDetailResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            source:
              type: string
            id:
              type: string
            type:
              type: string
            language:
              type: string
              nullable: true
            timestamp:
              type: string
              format: date-time
        data:
          $ref: '#/components/schemas/BookDetail'
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/BookSummary'
        - type: object
          properties:
            publishers:
              type: array
              items:
                type: string
            metadata:
              type: object
              description: Métadonnées additionnelles spécifiques au provider
              properties:
                places:
                  type: array
                  items:
                    type: string
                  description: Lieux mentionnés (Open Library)
                times:
                  type: array
                  items:
                    type: string
                  description: Périodes temporelles (Open Library)
                people:
                  type: array
                  items:
                    type: string
                  description: Personnages/personnes (Open Library)
                externalLinks:
                  type: array
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                      url:
                        type: string
                        format: uri
                format:
                  type: string
                  description: Format physique (Open Library)
                workId:
                  type: string
                  description: ID de l'œuvre parente (Open Library)
                maturityRating:
                  type: string
                  description: Classification d'âge (Google Books)
                previewLink:
                  type: string
                  format: uri
                  description: Lien de prévisualisation (Google Books)
                buyLink:
                  type: string
                  format: uri
                  description: Lien d'achat (Google Books)
                averageRating:
                  type: number
                  description: Note moyenne (Google Books)
                ratingsCount:
                  type: integer
                  description: Nombre de notes (Google Books)
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        timestamp:
          type: string
          format: date-time